Regroup theme views/modifiers
