+++ Passepartout/Library/Sources/AppUI/Theme/Theme.swift
@@ -81,171 +81,3 @@ public final class Theme: ObservableObject {
         animationCategories.contains(category) ? animation : nil
     }
 }
-
-// MARK: - Modifiers
-
-extension View {
-    public func themeModal<Content>(
-        isPresented: Binding<Bool>,
-        isRoot: Bool = false,
-        isInteractive: Bool = true,
-        content: @escaping () -> Content
-    ) -> some View where Content: View {
-        modifier(ThemeBooleanModalModifier(
-            isPresented: isPresented,
-            isRoot: isRoot,
-            isInteractive: isInteractive,
-            modal: content
-        ))
-    }
-
-    public func themeModal<Content, T>(
-        item: Binding<T?>,
-        isRoot: Bool = false,
-        isInteractive: Bool = true,
-        content: @escaping (T) -> Content
-    ) -> some View where Content: View, T: Identifiable {
-        modifier(ThemeItemModalModifier(
-            item: item,
-            isRoot: isRoot,
-            isInteractive: isInteractive,
-            modal: content
-        ))
-    }
-
-#if os(iOS)
-    public func themePopover<Content>(
-        isPresented: Binding<Bool>,
-        content: @escaping () -> Content
-    ) -> some View where Content: View {
-        modifier(ThemeBooleanPopoverModifier(
-            isPresented: isPresented,
-            popover: content
-        ))
-    }
-#endif
-
-    public func themeConfirmation(
-        isPresented: Binding<Bool>,
-        title: String,
-        isDestructive: Bool = false,
-        action: @escaping () -> Void
-    ) -> some View {
-        modifier(ThemeConfirmationModifier(
-            isPresented: isPresented,
-            title: title,
-            isDestructive: isDestructive,
-            action: action
-        ))
-    }
-
-    public func themeNavigationStack(if condition: Bool, closable: Bool = false, path: Binding<NavigationPath>) -> some View {
-        modifier(ThemeNavigationStackModifier(condition: condition, closable: closable, path: path))
-    }
-
-    public func themeForm() -> some View {
-        modifier(ThemeFormModifier())
-    }
-
-    public func themeManualInput() -> some View {
-        modifier(ThemeManualInputModifier())
-    }
-
-    public func themeSection(header: String? = nil, footer: String? = nil) -> some View {
-        modifier(ThemeSectionWithHeaderFooterModifier(header: header, footer: footer))
-    }
-
-    public func themeLockScreen() -> some View {
-#if !os(tvOS)
-        modifier(ThemeLockScreenModifier(lockedContent: LogoView.init))
-#else
-        self
-#endif
-    }
-
-#if !os(tvOS)
-    public func themeWindow(width: CGFloat, height: CGFloat) -> some View {
-        modifier(ThemeWindowModifier(size: .init(width: width, height: height)))
-    }
-
-    public func themeNavigationDetail() -> some View {
-        modifier(ThemeNavigationDetailModifier())
-    }
-
-    public func themePlainButton(action: @escaping () -> Void) -> some View {
-        modifier(ThemePlainButtonModifier(action: action))
-    }
-
-    @ViewBuilder
-    public func themeMultiLine(_ isMultiLine: Bool) -> some View {
-        if isMultiLine {
-            multilineTextAlignment(.leading)
-        } else {
-            themeTruncating()
-        }
-    }
-
-    public func themeTruncating(_ mode: Text.TruncationMode = .middle) -> some View {
-        lineLimit(1)
-            .truncationMode(mode)
-    }
-
-    public func themeEmptyMessage() -> some View {
-        modifier(ThemeEmptyMessageModifier())
-    }
-
-    public func themeError(_ isError: Bool) -> some View {
-        modifier(ThemeErrorModifier(isError: isError))
-    }
-
-    public func themeAnimation<T>(on value: T, category: ThemeAnimationCategory) -> some View where T: Equatable {
-        modifier(ThemeAnimationModifier(value: value, category: category))
-    }
-
-    public func themeTrailingValue(_ value: CustomStringConvertible?, truncationMode: Text.TruncationMode = .tail) -> some View {
-        modifier(ThemeTrailingValueModifier(value: value, truncationMode: truncationMode))
-    }
-
-    public func themeGridHeader(title: String?) -> some View {
-        modifier(ThemeGridSectionModifier(title: title))
-    }
-
-    public func themeGridCell(isSelected: Bool) -> some View {
-        modifier(ThemeGridCellModifier(isSelected: isSelected))
-    }
-
-    public func themeHoverListRow() -> some View {
-        modifier(ThemeHoverListRowModifier())
-    }
-
-    public func themeTip(_ text: String, edge: Edge) -> some View {
-        modifier(ThemeTipModifier(text: text, edge: edge))
-    }
-#endif
-}
-
-// MARK: - Views
-
-extension Theme {
-
-#if !os(tvOS)
-    public func listSection<ItemView: View, T: EditableValue>(
-        _ title: String,
-        addTitle: String,
-        originalItems: Binding<[T]>,
-        emptyValue: (() async -> T)? = nil,
-        @ViewBuilder itemLabel: @escaping (Bool, Binding<T>) -> ItemView
-    ) -> some View {
-        EditableListSection(
-            title,
-            addTitle: addTitle,
-            originalItems: originalItems,
-            emptyValue: emptyValue,
-            itemLabel: itemLabel,
-            removeLabel: ThemeEditableListSection.RemoveLabel.init(action:),
-            editLabel: ThemeEditableListSection.EditLabel.init
-        )
-    }
-#endif
-
-}
